routerdb_django/requirements.txt
-i https://pypi.python.org/simple
asgiref==3.7.2 ; python_version >= '3.7'
beautifulsoup4==4.12.2
certifi==2023.7.22 ; python_version >= '3.6'
charset-normalizer==3.2.0 ; python_full_version >= '3.7.0'
django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.4 ; python_version >= '3.5'
numpy==1.25.2 ; python_version >= '3.10'
packaging==23.1 ; python_version >= '3.7'
pandas==2.0.3
psycopg2-binary==2.9.7
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.4.1 ; python_version >= '3.7'
sqlparse==0.4.4 ; python_version >= '3.5'
typing-extensions==4.7.1 ; python_version < '3.11'
tzdata==2023.3 ; python_version >= '2'
urllib3==2.0.4 ; python_version >= '3.7'
